+++ ci.sh
@@ -1,24 +1,29 @@
 #!/usr/bin/env bash
 
+function build {
+    cargo build --release "$@"
+    cargo clippy --no-deps "$@"
+}
+
 set -ex
 
 echo Build with default features
-cargo build --release
+build
 
 echo Build with no default features
-cargo build --release --no-default-features
+build --no-default-features
 
 echo Build with only std feature
-cargo build --release --no-default-features --features std
+build --no-default-features --features std
 
 echo Build with only swash feature
-cargo build --release --no-default-features --features swash
+build --no-default-features --features swash
 
 echo Build with only syntect feature
-cargo build --release --no-default-features --features syntect
+build --no-default-features --features syntect
 
 echo Build with all features
-cargo build --release --all-features
+build --all-features
 
 echo Run tests
 cargo test
