+++ src/lib.rs
@@ -54,6 +54,25 @@
 //! });
 //! ```
 
+// Not interested in these lints
+#![allow(clippy::new_without_default)]
+
+// TODO: address ocurrances and then deny
+// Indexing a slice can cause panics and that is something we always want to avoid
+#![allow(clippy::indexing_slicing)]
+// Overflows can produce unpredictable results and are only checked in debug builds
+ #![allow(clippy::integer_arithmetic)]
+
+// Dereferencing unalinged pointers may be undefined behavior
+#![deny(clippy::cast_ptr_alignment)]
+// Avoid panicking in without information about the panic. Use expect
+#![deny(clippy::unwrap_used)]
+// This is usually a serious issue - a missing import of a define where it is interpreted
+// as a catch-all variable in a match, for example
+#![deny(unreachable_patterns)]
+// Ensure that all must_use results are used
+#![deny(unused_must_use)]
+
 #![cfg_attr(not(feature = "std"), no_std)]
 
 extern crate alloc;
